modularizando multer, y terminando cookies
