+++ myapp/src/middlewares/multer/products.js
@@ -1,0 +1,22 @@
+const multer = require('multer')
+const storage = multer.diskStorage({
+    destination: function (req, file, cb) {
+    cb(null, path.join('public/imgs/products'))
+  },
+  filename: function (req, file, cb) {
+    cb(null, Date.now() + file.originalname);
+  }
+})
+
+module.exports = multer({ 
+    storage,
+    fileFilter: (req, file, cb) => {
+        const acceptExt = ['.jpg', '.png', '.webp', '.jpeg'];
+        const ext = path.extname(file.originalname);
+        
+        if (!acceptExt.includes(ext)) {
+            req.files = file;
+        }
+        cb(null, acceptExt.includes(ext));
+    }
+})
