+++ myapp/src/middlewares/setLocals.js
@@ -3,7 +3,6 @@ module.exports = (req, res, next) => {
     if (!req.session.user) {
         return next();
     }
-
     res.locals.userLog = req.session.user;
 
     return next();
