+++ myapp/src/routes/users.js
@@ -6,8 +6,8 @@ const usersController = require('../controllers/usersController')
 
 /* GET Users page. */
 
-router.post('/login', validator.login, usersController.processLogin);
 router.get('/:id', usersController.index);
+router.post('/login', validator.login, usersController.processLogin);
 router.get('/register', usersController.register);
 router.post('/register', validator.register, usersController.createUser);
 router.post('/register-business', usersController.createBusinessUser);
